pyproject.toml
[tool.poetry]
name = "repl-nix-workio-website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
sqlalchemy = "^2.0.32"
consultor = "^0.2.0"
pyjwt = "^2.9.0"
flask-jwt-extended = "^4.6.0"
werkzeug = "^3.0.4"
flask-login = "^0.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
